cli/src/aps_cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aps-cli"
version = "0.1.0"
description = "APS CLI"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
authors = [{ name = "APS Contributors" }]

dependencies = [
  "PyYAML>=6.0",
  "requests>=2.31",
  "cryptography>=41",
]

[project.scripts]
aps = "aps_cli.app:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["aps_cli*"]

[tool.setuptools]
package-dir = {"" = "src"}


